Refactor, uses MFA Cosigner example rather than test example
