+++ examples/mfa/example.go
@@ -2,28 +2,16 @@ package main
 
 import (
 	"context"
-	"encoding/json"
 	"fmt"
 	"os"
 
 	"github.com/lestrrat-go/jwx/v2/jwa"
 	"github.com/openpubkey/openpubkey/client"
 	"github.com/openpubkey/openpubkey/client/providers"
-	"github.com/openpubkey/openpubkey/examples/mfa/webauthn"
+	"github.com/openpubkey/openpubkey/cosigner/cosclient"
+	"github.com/openpubkey/openpubkey/examples/mfa/mfacosigner"
 	"github.com/openpubkey/openpubkey/util"
 )
-
-// Variables for building our google provider
-// var (
-// 	clientID = "184968138938-g1fddl5tglo7mnlbdak8hbsqhhf79f32.apps.googleusercontent.com"
-// 	// The clientSecret was intentionally checked in for the purposes of this example,. It holds no power. Do not report as a security issue
-// 	clientSecret = "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" // Google requires a ClientSecret even if this a public OIDC App
-// 	issuer       = "https://accounts.google.com"
-// 	scopes       = []string{"openid profile email"}
-// 	redirURIPort = "3000"
-// 	callbackPath = "/login-callback"
-// 	redirectURI  = fmt.Sprintf("http://localhost:%v%v", redirURIPort, callbackPath)
-// )
 
 var (
 	clientID = "115045953232-9e3co450v3dagfh5q8pplip4otb5bgim.apps.googleusercontent.com"
@@ -36,12 +24,66 @@ var (
 	redirectURI  = fmt.Sprintf("http://localhost:%v%v", redirURIPort, callbackPath)
 )
 
+// func main() {
+// 	clientKey, err := util.GenKeyPair(jwa.ES256)
+// 	if err != nil {
+// 		fmt.Println("error generating key pair:", err)
+// 		return
+// 	}
+
+// 	provider := &providers.GoogleOp{
+// 		ClientID:     clientID,
+// 		ClientSecret: clientSecret,
+// 		Issuer:       issuer,
+// 		Scopes:       scopes,
+// 		RedirURIPort: redirURIPort,
+// 		CallbackPath: callbackPath,
+// 		RedirectURI:  redirectURI,
+// 	}
+
+// 	command := os.Args[1]
+// 	switch command {
+// 	case "login":
+// 		opk := &client.OpkClient{
+// 			Op: provider,
+// 		}
+
+// 		pkt, err := opk.OidcAuth(context.TODO(), clientKey, jwa.ES256, map[string]any{}, false)
+// 		if err != nil {
+// 			fmt.Println("error generating key pair: ", err)
+// 			return
+// 		}
+// 		fmt.Println("New PK token generated")
+
+// 		// Verify our pktoken including the cosigner signature
+// 		if err := client.VerifyPKToken(context.TODO(), pkt, provider); err != nil {
+// 			fmt.Println("failed to verify PK token:", err)
+// 		}
+
+// 		pktJson, err := json.Marshal(pkt)
+// 		if err != nil {
+// 			fmt.Println("error serializing pktJson: ", err)
+// 			return
+// 		}
+// 		fmt.Println(string(pktJson))
+
+// 	case "mfa":
+// 		rpID := "localhost"
+// 		serverUri := "http://localhost:3003"
+// 		rpOrigin := "http://localhost:3003"
+// 		rpDisplayName := "OpenPubkey"
+// 		_, err := webauthn.New(serverUri, rpID, rpOrigin, rpDisplayName)
+// 		if err != nil {
+// 			fmt.Println("error starting mfa server: ", err)
+// 			return
+// 		}
+
+// 	default:
+// 		fmt.Println("Unrecognized command:", command)
+// 	}
+// }
+
 func main() {
-	clientKey, err := util.GenKeyPair(jwa.ES256)
-	if err != nil {
-		fmt.Println("error generating key pair:", err)
-		return
-	}
 
 	provider := &providers.GoogleOp{
 		ClientID:     clientID,
@@ -53,14 +95,30 @@ func main() {
 		RedirectURI:  redirectURI,
 	}
 
+	cosClient := &cosclient.AuthCosignerClient{
+		Issuer:      "http://localhost:3003",
+		RedirectURI: "http://localhost:3000/mfacallback",
+	}
+	if len(os.Args) < 2 {
+		fmt.Printf("Example MFA Cosigner: command choices are: login, mfa")
+		return
+	}
+
 	command := os.Args[1]
 	switch command {
 	case "login":
 		opk := &client.OpkClient{
-			Op: provider,
+			Op:     provider,
+			MfaCos: cosClient,
 		}
 
-		pkt, err := opk.OidcAuth(context.TODO(), clientKey, jwa.ES256, map[string]any{}, false)
+		clientKey, err := util.GenKeyPair(jwa.ES256)
+		if err != nil {
+			fmt.Println("error generating key pair:", err)
+			return
+		}
+
+		pkt, err := opk.CosAuth(context.TODO(), clientKey, jwa.ES256, map[string]any{}, false)
 		if err != nil {
 			fmt.Println("error generating key pair: ", err)
 			return
@@ -68,23 +126,23 @@ func main() {
 		fmt.Println("New PK token generated")
 
 		// Verify our pktoken including the cosigner signature
+		// TODO: Use this to test cosigner signature issuer allow list
 		if err := client.VerifyPKToken(context.TODO(), pkt, provider); err != nil {
 			fmt.Println("failed to verify PK token:", err)
 		}
-
-		pktJson, err := json.Marshal(pkt)
-		if err != nil {
-			fmt.Println("error serializing pktJson: ", err)
-			return
+		if err := pkt.VerifyCosignerSignature(); err != nil {
+			fmt.Println("failed to verify PK token cosigner signature:", err)
+		} else {
+			fmt.Println("Verified PK token cosigner signature!")
 		}
-		fmt.Println(string(pktJson))
 
+		os.Exit(0)
 	case "mfa":
 		rpID := "localhost"
 		serverUri := "http://localhost:3003"
 		rpOrigin := "http://localhost:3003"
 		rpDisplayName := "OpenPubkey"
-		_, err := webauthn.New(serverUri, rpID, rpOrigin, rpDisplayName)
+		_, err := mfacosigner.New(serverUri, rpID, rpOrigin, rpDisplayName)
 		if err != nil {
 			fmt.Println("error starting mfa server: ", err)
 			return
@@ -92,6 +150,7 @@ func main() {
 
 	default:
 		fmt.Println("Unrecognized command:", command)
+		fmt.Printf("Example MFA Cosigner: command choices are: login, mfa")
 	}
 
 }
