+++ examples/mfa/mfacosigner/mfacosigner_test.go
@@ -1,4 +1,4 @@
-package webauthn
+package mfacosigner
 
 import (
 	"crypto"
