--- examples/mfa/webauthn/static/index.html
@@ -1,183 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-</head>
-<body>
-    <div id="status">Authenticating with MFA...</div>
-
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <script>
-        // Convert a Base64 URL-encoded string to an ArrayBuffer
-        function base64UrlDecode(value) {
-            let avlue = value.replace(/\-/g, '+') // Replace '-' with '+'
-                .replace(/\_/g, '/') // Replace '_' with '/'
-                .replace(/=+$/, ''); // Remove padding
-            return Uint8Array.from(atob(avlue), c => c.charCodeAt(0));
-        }
-
-        // Convert an ArrayBuffer to Base64 URL-encoded string
-        function base64UrlEncode(buffer) {
-            let binary = '';
-            let bytes = new Uint8Array(buffer);
-            let len = bytes.byteLength;
-            for (let i = 0; i < len; i++) {
-                binary += String.fromCharCode(bytes[i]);
-            }
-
-            return window.btoa(binary)
-                .replace(/\+/g, '-') // Replace '+' with '-'
-                .replace(/\//g, '_') // Replace '/' with '_'
-                .replace(/=+$/, ''); // Remove padding
-        }
-
-        async function register(authid) {
-            let response = await fetch(`/register/begin?authid=${authid}`, {
-                method: 'GET'
-            });
-
-            if (!response.ok) {
-                throw new Error(`HTTP error! status: ${response.status}`);
-            }
-
-            let data = await response.json();
-            data.publicKey.user.id = base64UrlDecode(data.publicKey.user.id);
-            data.publicKey.challenge = base64UrlDecode(data.publicKey.challenge)
-
-            credentials = await navigator.credentials.create(data);
-            let credentialsForServer = {
-                id: credentials.id,
-                type: credentials.type,
-                rawId: base64UrlEncode(credentials.rawId),
-                response: {
-                    clientDataJSON: base64UrlEncode(credentials.response.clientDataJSON),
-                    attestationObject: base64UrlEncode(credentials.response.attestationObject)
-                }
-            };
-            
-
-            // TODO: Is this needed? Where is cred used? Is credentials.response.userHandle always false so we never hit this?
-            if (credentials.response.userHandle) {
-                cred.response.userHandle = base64UrlEncode(credentials.response.userHandle);
-            }
-
-            // Send credentials back to server to save
-            let res = await fetch(`/register/finish?authid=${authid}`, {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json'
-                },
-                body: JSON.stringify(credentialsForServer)
-            });
-
-            if (!res.ok) {
-                throw new Error(`HTTP error! status: ${res.status}`);
-            }
-
-            console.log("Registration complete!");
-        }
-
-        async function login(authid) {
-            let response = await fetch(`/login/begin?authid=${authid}`, {
-                method: 'GET'
-            });
-
-            if (!response.ok) {
-                throw new Error(`HTTP error! status: ${response.status}`);
-            }
-
-            let data = await response.json();
-            data.publicKey.challenge = base64UrlDecode(data.publicKey.challenge)
-            data.publicKey.allowCredentials.forEach(function (listItem) {
-                listItem.id = base64UrlDecode(listItem.id)
-            });
-
-            let assertion = await navigator.credentials.get(data);
-
-            // Send credentials back to server to save
-            let res = await fetch(`/login/finish?authid=${authid}`, {
-                method: 'POST',
-                headers: {
-                    'Content-Type': 'application/json'
-                },
-                body: JSON.stringify({
-                    id: assertion.id,
-                    rawId: base64UrlEncode(assertion.rawId),
-                    type: assertion.type,
-                    response: {
-                        authenticatorData: base64UrlEncode(assertion.response.authenticatorData),
-                        clientDataJSON: base64UrlEncode(assertion.response.clientDataJSON),
-                        signature: base64UrlEncode(assertion.response.signature),
-                        userHandle: base64UrlEncode(assertion.response.userHandle),
-                    }
-                })
-            });
-
-            if (!res.ok) {
-                throw new Error(`HTTP error! status: ${res.status}`);
-            }
-
-            let authcodeData = await res.json();
-            window.location.replace(authcodeData.redirect_uri);
-
-            console.log("Login complete!");
-        }
-
-        document.addEventListener('DOMContentLoaded', async (event) => {
-            const statusDiv = document.getElementById('status');
-            
-            
-            const queryString = window.location.search;
-            let searchParams = new URLSearchParams(queryString);
-            const authid = searchParams.get('authid')
-            console.log(authid);
-
-            const response = await fetch(`/check-registration?authid=${authid}`, {
-                method: 'GET',
-            });
-
-
-            let data = await response.json();
-
-            if (!data.isRegistered) {
-                try {
-                    statusDiv.textContent = "Registering user with WebAuthn...";
-                    await register(authid);
-                } catch(error) {
-                    statusDiv.textContent = `Error during registration: ${error}`;
-                    await fetch('/done', {
-                        method: 'POST',
-                        headers: {
-                            'Content-Type': 'application/json'
-                        },
-                        body: `Error during registration: ${error}`
-                    });
-                }
-            }
-
-            try {
-                statusDiv.textContent = "Authenticating user with WebAuthn...";
-                let response = await login(authid);
-            } catch(error) {
-                statusDiv.textContent = `Error during login: ${error}`;
-                await fetch('/done', {
-                    method: 'POST',
-                    headers: {
-                        'Content-Type': 'application/json'
-                    },
-                    body: `Error during login: ${error}`
-                });
-            }
-
-
-
-            statusDiv.textContent = "MFA authentication complete!";
-            await fetch('/done', {
-                method: 'GET',
-            });
-
-  
-        })
-    </script>
-</body>
-</html>
